Separating utilities into their own package. Adding Map chooseOneFrom methods Renamed Collection chooseOneFrom method Added example usage
